code/Dicom Management system/requirements.txt
django~=2.2

pydicom~=2.1.2
fhirpy~=1.2.1
py~=1.10.0
colorama~=0.4.4
pip~=21.1.2
lxml~=4.6.3
attrs~=20.3.0
pytz~=2020.1
toml~=0.10.2
numpy~=1.19.4
Pillow~=8.2.0
sqlparse~=0.3.1
mysqlclient~=2.0.3
PyYAML~=5.3.1
Jinja2~=2.11.2
pytest~=6.2.4
wcwidth~=0.2.5
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.0
pyparsing~=2.4.7
setuptools~=41.2.0